backend/requirements.txt
flask
pymongo
dnspython
python-dotenv
jsonify
jinja2
fastapi
pydantic